+++ tests/dummy/app/controllers/components-examples/flexberry-simpleolv/limit-function-example.js
@@ -1,0 +1,4 @@
+import ListFormController from 'ember-flexberry/controllers/list-form';
+
+export default ListFormController.extend({
+});